fixed params in load dataset and renamed _core.py to _signal.py
