+++ vessim/_signal.py
